% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{stdcols}
\alias{stdcols}
\title{Standard colors in red, blue, orange}
\format{
An object of class \code{list} of length 3.
}
\usage{
stdcols
}
\description{
Standard colors in red, blue, orange
}
\keyword{datasets}
